@@ -1,104 +1,94 @@
-import 'package:app_cantina/helpers/cliente_helper.dart';
+//import 'package:app_cantina/helpers/cliente_helper.dart';
 import 'package:flutter/material.dart';
 
 class HomePage extends StatefulWidget {
-  const HomePage({Key key}) : super(key: key);
+  const HomePage({Key? key}) : super(key: key);
 
   @override
   _HomePageState createState() => _HomePageState();
 }
 
 class _HomePageState extends State<HomePage> {
-  ClienteHelper helper = ClienteHelper();
+  final _formKey = GlobalKey<FormState>();
+  GlobalKey<FormState> _formKeyCad = new GlobalKey();
+  // ClienteHelper helper = ClienteHelper();
 
   @override
   void initState() {
     super.initState();
   }
 
-  Widget LoginPage() {
+  Widget loginPage() {
     const primaryColor = Color(0xFFFA4A0C);
     return SingleChildScrollView(
-      padding: const EdgeInsets.all(50.0),
-      child: Column(
-        children: <Widget>[
-          TextFormField(
-            decoration: const InputDecoration(
-              labelText: 'E-mail',
-              labelStyle: TextStyle(color: Colors.grey),
-              enabledBorder: UnderlineInputBorder(
-                borderSide: BorderSide(color: Colors.grey),
-              ),
-            ),
-            validator: (String value) {
-              if (value == null || value.isEmpty) {
-                return 'Por favor preencha seu e-mail';
-              }
-              return null;
-            },
-          ),
-          const Divider(color: Colors.transparent),
-          TextFormField(
-            obscureText: true,
-            decoration: const InputDecoration(
-              labelText: 'Senha',
-              labelStyle: TextStyle(color: Colors.grey),
-              enabledBorder: UnderlineInputBorder(
-                  borderSide: BorderSide(color: Colors.grey)),
-            ),
-            validator: (String value) {
-              if (value == null || value.isEmpty) {
-                return 'Por favor preencha a senha';
-              }
-              return null;
-            },
-          ),
-          const Divider(
-            height: 35,
-            color: Colors.transparent,
-          ),
-          const Text('Recuperar Senha',
-              style: TextStyle(
-                  color: primaryColor,
-                  fontSize: 14,
-                  fontWeight: FontWeight.bold),
-              textAlign: TextAlign.left),
-          const Divider(
-            height: 30,
-            color: Colors.transparent,
-          ),
-          ElevatedButton(
-            child: const Text('Logar',
-                style: TextStyle(
-                    color: Colors.white, fontWeight: FontWeight.bold)),
-            style: ElevatedButton.styleFrom(
-              primary: primaryColor,
-              padding: const EdgeInsets.fromLTRB(130, 20, 130, 20),
-              shape: RoundedRectangleBorder(
-                borderRadius: BorderRadius.circular(20),
-              ),
-            ),
-            onPressed: () {
-              Navigator.pushNamed(context, '/home');
-            },
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget CadastroPage() {
-    return Container(
-      child: Center(
+      child: Padding(
+        padding: const EdgeInsets.fromLTRB(30, 10, 30, 10),
         child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: const <Widget>[
-            Text(
-              "Cadastro",
-              style: TextStyle(fontSize: 20),
-            ),
-            Padding(
-              padding: EdgeInsets.only(top: 20),
+          children: [
+            Form(
+              key: _formKey,
+              child: Column(
+                children: <Widget>[
+                  TextFormField(
+                    decoration: const InputDecoration(
+                      labelText: 'E-mail',
+                      labelStyle: TextStyle(color: Colors.grey),
+                      enabledBorder: UnderlineInputBorder(
+                        borderSide: BorderSide(color: Colors.grey),
+                      ),
+                    ),
+                    validator: (value) {
+                      if (value == null || value.isEmpty) {
+                        return 'Por favor preencha seu e-mail';
+                      }
+                      return null;
+                    },
+                  ),
+                  const Divider(color: Colors.transparent),
+                  TextFormField(
+                    obscureText: true,
+                    decoration: const InputDecoration(
+                      labelText: 'Senha',
+                      labelStyle: TextStyle(color: Colors.grey),
+                      enabledBorder: UnderlineInputBorder(
+                          borderSide: BorderSide(color: Colors.grey)),
+                    ),
+                    validator: (value) {
+                      if (value == null || value.isEmpty) {
+                        return 'Por favor preencha a senha';
+                      }
+                      return null;
+                    },
+                  ),
+                  const Divider(
+                    height: 35,
+                    color: Colors.transparent,
+                  ),
+                  const Text('Recuperar Senha',
+                      style: TextStyle(
+                          color: primaryColor,
+                          fontSize: 14,
+                          fontWeight: FontWeight.bold),
+                      textAlign: TextAlign.left),
+                  const Divider(
+                    height: 30,
+                    color: Colors.transparent,
+                  ),
+                  ElevatedButton(
+                    child: const Text('Logar',
+                        style: TextStyle(
+                            color: Colors.white, fontWeight: FontWeight.bold)),
+                    style: ElevatedButton.styleFrom(
+                      primary: primaryColor,
+                      padding: const EdgeInsets.fromLTRB(100, 20, 100, 20),
+                      shape: RoundedRectangleBorder(
+                        borderRadius: BorderRadius.circular(20),
+                      ),
+                    ),
+                    onPressed: _sendForm,
+                  ),
+                ],
+              ),
             ),
           ],
         ),
@@ -106,15 +96,129 @@ class _HomePageState extends State<HomePage> {
     );
   }
 
+  _sendForm() {
+    if (_formKey.currentState!.validate()) {
+      Navigator.pushNamed(context, '/produtos_list');
+    }
+  }
+
+  Widget cadastroPage() {
+    const primaryColor = Color(0xFFFA4A0C);
+    return SingleChildScrollView(
+      child: Padding(
+        padding: const EdgeInsets.fromLTRB(30, 10, 30, 10),
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: <Widget>[
+            Form(
+              key: _formKeyCad,
+              child: Column(
+                children: <Widget>[
+                  TextFormField(
+                    decoration: const InputDecoration(
+                      labelText: 'Nome',
+                      labelStyle: TextStyle(color: Colors.grey),
+                      enabledBorder: UnderlineInputBorder(
+                        borderSide: BorderSide(color: Colors.grey),
+                      ),
+                    ),
+                    validator: (arg) {
+                      if (arg == null || arg.isEmpty) {
+                        return 'Preencha um nome';
+                      } else
+                        return null;
+                    },
+                  ),
+                  TextFormField(
+                    decoration: const InputDecoration(
+                      labelText: 'E-mail',
+                      labelStyle: TextStyle(color: Colors.grey),
+                      enabledBorder: UnderlineInputBorder(
+                        borderSide: BorderSide(color: Colors.grey),
+                      ),
+                    ),
+                    validator: (value) {
+                      if (value == null || value.isEmpty) {
+                        return 'Por favor preencha seu e-mail';
+                      }
+                      return null;
+                    },
+                  ),
+                  TextFormField(
+                    obscureText: true,
+                    decoration: const InputDecoration(
+                      labelText: 'Senha',
+                      labelStyle: TextStyle(color: Colors.grey),
+                      enabledBorder: UnderlineInputBorder(
+                        borderSide: BorderSide(color: Colors.grey),
+                      ),
+                    ),
+                    validator: (value) {
+                      if (value == null || value.isEmpty) {
+                        return 'Por favor preencha a senha';
+                      }
+                      return null;
+                    },
+                  ),
+                  TextFormField(
+                    obscureText: true,
+                    decoration: const InputDecoration(
+                      labelText: 'Confirmar Senha',
+                      labelStyle: TextStyle(color: Colors.grey),
+                      enabledBorder: UnderlineInputBorder(
+                        borderSide: BorderSide(color: Colors.grey),
+                      ),
+                    ),
+                    validator: (value) {
+                      if (value == null || value.isEmpty) {
+                        return 'Por favor preencha a senha';
+                      }
+                      return null;
+                    },
+                  ),
+                  const Divider(
+                    height: 35,
+                    color: Colors.transparent,
+                  ),
+                  ElevatedButton(
+                    child: const Text('Cadastrar',
+                        style: TextStyle(
+                            color: Colors.white, fontWeight: FontWeight.bold)),
+                    style: ElevatedButton.styleFrom(
+                      primary: primaryColor,
+                      padding: const EdgeInsets.fromLTRB(90, 20, 90, 20),
+                      shape: RoundedRectangleBorder(
+                        borderRadius: BorderRadius.circular(20),
+                      ),
+                    ),
+                    onPressed: _sendFormCad,
+                  ),
+                ],
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+
+  _sendFormCad() {
+    if (_formKey.currentState!.validate()) {
+      Navigator.pushNamed(context, '/cadatrar');
+    }
+  }
+
   Widget _buildImageRow() => Expanded(
         flex: 3,
         child: Container(
           alignment: Alignment.center,
-          child: Image.asset(
-            'assets/logo.png',
-            //fit: BoxFit.scaleDown,
-            height: 180,
-            alignment: Alignment.center,
+          child: SingleChildScrollView(
+            child: Image.asset(
+              'assets/logo.png',
+              //fit: BoxFit.scaleDown,
+              height: 180,
+              alignment: Alignment.center,
+            ),
           ),
         ),
       );
@@ -133,7 +237,7 @@ class _HomePageState extends State<HomePage> {
     );
   }
 
-  Widget _buildTextRow(Color primaryColor) => Expanded(
+  Widget _buildTextRow() => Expanded(
         flex: 7,
         child: DefaultTabController(
           length: 2,
@@ -155,7 +259,7 @@ class _HomePageState extends State<HomePage> {
                   indicator: BoxDecoration(
                     border: Border(
                       bottom: BorderSide(
-                        color: primaryColor,
+                        color: Color(0xFFFA4A0C),
                         width: 3,
                       ),
                     ),
@@ -169,8 +273,8 @@ class _HomePageState extends State<HomePage> {
             ),
             body: TabBarView(
               children: [
-                LoginPage(),
-                CadastroPage(),
+                loginPage(),
+                cadastroPage(),
               ],
             ),
           ),
@@ -181,63 +285,70 @@ class _HomePageState extends State<HomePage> {
   Widget build(BuildContext context) {
     const color = Colors.white;
     const primaryColor = Color(0xFFFA4A0C);
-    return //Scaffold(
-        /*bottomSheet: Container(
-        height: 50,
-        color: Colors.white,
-         child: Row(
-          mainAxisAlignment: MainAxisAlignment.spaceBetween,
-          children: [
-            Container(
-              height: 50,
-              width: 50,
-              child: Icon(
-                Icons.home,
-                color: primaryColor,
-              ),
-            ),
-            Container(
-              height: 50,
-              width: 50,
-              child: Icon(
-                Icons.search,
-                color: primaryColor,
-              ),
-            ),
-            Container(
-              height: 50,
-              width: 50,
-              child: Icon(
-                Icons.shopping_cart,
-                color: primaryColor,
-              ),
-            ),
-            Container(
-              height: 50,
-              width: 50,
-              child: Icon(
-                Icons.person,
-                color: primaryColor,
-              ),
-            ),
-          ],
-        ), 
-      ),*/
+    return
+
+        // Scaffold(
+        //     bottomSheet: Container(
+        //       height: 50,
+        //       color: Colors.white,
+        //       child: Row(
+        //         mainAxisAlignment: MainAxisAlignment.spaceBetween,
+        //         children: [
+        //           Container(
+        //             height: 50,
+        //             width: 50,
+        //             child: Icon(
+        //               Icons.home,
+        //               color: primaryColor,
+        //             ),
+        //           ),
+        //           Container(
+        //             height: 50,
+        //             width: 50,
+        //             child: Icon(
+        //               Icons.search,
+        //               color: primaryColor,
+        //             ),
+        //           ),
+        //           Container(
+        //             height: 50,
+        //             width: 50,
+        //             child: Icon(
+        //               Icons.shopping_cart,
+        //               color: primaryColor,
+        //             ),
+        //           ),
+        //           Container(
+        //             height: 50,
+        //             width: 50,
+        //             child: Icon(
+        //               Icons.person,
+        //               color: primaryColor,
+        //             ),
+        //           ),
+        //         ],
+        //       ),
+        //     ),
         Container(
-            padding: const EdgeInsets.only(top: 50),
-            height: MediaQuery.of(context).size.height,
-            width: MediaQuery.of(context).size.width,
-            decoration: const BoxDecoration(
-              color: color,
-              borderRadius: BorderRadius.only(
-                  bottomLeft: Radius.circular(30.0),
-                  bottomRight: Radius.circular(30.0)),
-            ),
-            child: Column(
-              children: [
-                _buildImageRow(),
-                _buildTextRow(primaryColor),
-              ],
-            ));
+      padding: const EdgeInsets.only(top: 50),
+      height: MediaQuery.of(context).size.height,
+      width: MediaQuery.of(context).size.width,
+      decoration: const BoxDecoration(
+        color: color,
+        borderRadius: BorderRadius.only(
+            bottomLeft: Radius.circular(30.0),
+            bottomRight: Radius.circular(30.0)),
+      ),
+      child: Column(
+        children: [
+          Expanded(
+            child: Column(children: [
+              _buildImageRow(),
+              _buildTextRow(),
+            ]),
+          ),
+        ],
+      ),
+    );
   }
 }
